requirements.txt
mysql-connector-python==8.2.0
pandas==2.0.3
openpyxl==3.1.2
sqlalchemy==2.0.23
pytest-order==1.1.0
PyYAML==6.0.1